+++ packages/@livestore/livestore/src/index.ts
@@ -45,7 +45,6 @@ export * from '@livestore/common/schema'
 export {
   sql,
   SessionIdSymbol,
-  type BootDb,
   type BootStatus,
   type SynchronousDatabase,
   type DebugInfo,