+++ packages/@livestore/react/src/LiveStoreProvider.test.tsx
@@ -1,6 +1,5 @@
-import type { BootDb } from '@livestore/common'
 import { sql } from '@livestore/common'
-import { querySQL } from '@livestore/livestore'
+import { querySQL, type Store } from '@livestore/livestore'
 import { Schema } from '@livestore/utils/effect'
 import { makeInMemoryAdapter } from '@livestore/web'
 import { render, screen, waitFor, waitForElementToBeRemoved } from '@testing-library/react'
@@ -30,8 +29,8 @@ describe('LiveStoreProvider', () => {
 
     const Root = ({ forceUpdate }: { forceUpdate: number }) => {
       const bootCb = React.useCallback(
-        (db: BootDb) =>
-          db.execute(sql`INSERT OR IGNORE INTO todos (id, text, completed) VALUES ('t1', 'buy milk', 0);`),
+        (store: Store) =>
+          store.__execute(sql`INSERT OR IGNORE INTO todos (id, text, completed) VALUES ('t1', 'buy milk', 0);`),
         [],
       )
       // eslint-disable-next-line react-hooks/exhaustive-deps
@@ -81,8 +80,8 @@ describe('LiveStoreProvider', () => {
 
     const Root = ({ forceUpdate }: { forceUpdate: number }) => {
       const bootCb = React.useCallback(
-        (db: BootDb) =>
-          db.execute(sql`INSERT INTO todos_mising_table (id, text, completed) VALUES ('t1', 'buy milk', 0);`),
+        (store: Store) =>
+          store.__execute(sql`INSERT INTO todos_mising_table (id, text, completed) VALUES ('t1', 'buy milk', 0);`),
         [],
       )
       // eslint-disable-next-line react-hooks/exhaustive-deps