+++ CHANGELOG.md
@@ -1,8 +1,9 @@
 # Changelog
 
 > NOTE: LiveStore is still in alpha and releases can include breaking changes. See [state of the project](https://preview.livestore.dev/reference/state-of-the-project/) for more info.
+> LiveStore is following a semver-like release strategy where breaking changes are released in minor versions before the 1.0 release.
 
-## 0.0.58
+## 0.1.0
 
 - Prepared the foundations for the upcoming [rebase sync protocol](https://github.com/livestorejs/livestore/issues/195)
   - New event id strategy (uses a global event id integer sequence number)
@@ -13,6 +14,18 @@
   - If you're using `effect` in your project, make sure to install version `3.10.x`
 		- Note the new version of `effect` now includes `Schema` directly, so `@effect/schema` is no longer needed as a separate dependency. (See [Effect blog post](https://effect.website/blog/releases/effect/310/#effectschema-moved-to-effectschema).)
 - Breaking: Moved `effect-db-schema` to `@livestore/db-schema` (mostly an internal change unless you're using the package directly)
+
+- Breaking: Adjusted `boot` signature when creating a store to now pass in a `Store` instead of a helper database object
+  ```tsx
+	<LiveStoreProvider
+		schema={schema}
+		boot={(store) => store.mutate(mutations.addTodo({ id: cuid(), text: 'Make coffee' }))}
+		adapter={adapter}
+		batchUpdates={batchUpdates}
+	>
+		// ...
+	</LiveStoreProvider>
+	```
 
 ### React integration